# api/src/middleware/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Events API
  description: An API for events management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 0.1.0
servers:
  - url: localhost:5000
paths:
  /:
    get:
      summary: Return available resources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Root'
  /events:
    get:
      summary: Return a list of events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Events'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      summary: Add a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostEvent'
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/BadRequest'
  /events/{eventId}:
    get:
      summary: Returns a single event
      parameters:
        - in: path
          name: eventId
          schema:
            type: integer
          required: true
          description: Numeric ID of the event to get
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Event:
      description: OK
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/Event'
  schemas:
    Root:
      type: object
    PostEvent:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        date:
          type: string
      required:
        - name
        - description
        - date
    Event:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        date:
          type: string
      required:
        - id
        - name
        - description
        - date
    Events:
      type: object
      properties:
        data:
          type: array
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
          required:
            - message